Moving eventsummarylistitem into composables module.
