+++ composables/build.gradle.kts
@@ -23,6 +23,7 @@ kotlin {
         val commonMain by getting {
             dependencies {
                 implementation(project(":shared"))
+                implementation(project(":core-displaymodels"))
                 implementation(compose.foundation)
                 implementation(compose.ui)
                 implementation(compose("org.jetbrains.compose.ui:ui-tooling"))
